refactor(components): extract PokedexCard from PokedexGrid
Extracted the single card logic out of the PokedexGrid component to improve maintainability and readability. The complex rendering of the Pokedex card is now safely housed in PokedexCard.tsx, making the PokedexGrid.tsx much shorter and easier to reason about.
